requirements.txt
paho-mqtt==1.6.1
python-dotenv==0.20.0
simple-websocket-server==0.4.1
websockets==10.3
